Added a messaging system. Creating a save and creation system to make levels easier to make in the long run.

+++ Assets/Scripts/UI/UIManager.cs
@@ -27,7 +27,7 @@ namespace AtomicKitchenChaos.UI {
             exoticMaterialPanelUI.AddExoticMaterialUI(material, count);
         }
 
-        public void SetQuarkCount(int quarkCount) {
+        public void SetQuarkCount(long quarkCount) {
             hudUI.SetQuarkCount(quarkCount);
         }
     }
